requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_google_genai==2.1.3
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.40
